easybatch-extensions/easybatch-yaml/src/test/java/org/easybatch/extensions/yaml/YamlRecordMapperTest.java
/**
 * The MIT License
 *
 *   Copyright (c) 2020, Mahmoud Ben Hassine (mahmoud.benhassine@icloud.com)
 *
 *   Permission is hereby granted, free of charge, to any person obtaining a copy
 *   of this software and associated documentation files (the "Software"), to deal
 *   in the Software without restriction, including without limitation the rights
 *   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 *   copies of the Software, and to permit persons to whom the Software is
 *   furnished to do so, subject to the following conditions:
 *
 *   The above copyright notice and this permission notice shall be included in
 *   all copies or substantial portions of the Software.
 *
 *   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 *   IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 *   FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 *   AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 *   LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 *   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 *   THE SOFTWARE.
 */
package org.easybatch.extensions.yaml;

import org.easybatch.core.record.Record;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

import static org.assertj.core.api.Assertions.assertThat;
import static org.easybatch.core.util.Utils.LINE_SEPARATOR;
import static org.mockito.Mockito.when;

@RunWith(MockitoJUnitRunner.class)
public class YamlRecordMapperTest {

    @Mock
    private YamlRecord record;

    private YamlRecordMapper<Contact> recordMapper;

    @Before
    public void setUp() throws Exception {
        recordMapper = new YamlRecordMapper<>(Contact.class);
        when(record.getPayload()).thenReturn("name: Foo" + LINE_SEPARATOR + "age: 28");
    }

    @Test
    public void processRecord() throws Exception {
        Record<Contact> actual = recordMapper.processRecord(record);

        assertThat(actual).isNotNull();
        Contact contact = actual.getPayload();

        assertThat(contact).isNotNull();
        assertThat(contact.getName()).isEqualTo("Foo");
        assertThat(contact.getAge()).isEqualTo(28);
    }

}
